init android

+++ util/src/main/java/fr/ensibs/util/graphic/ISnapshotLayer.java
@@ -36,14 +36,14 @@ public interface ISnapshotLayer<I extends IImage>
      *
      * @return the width of the area used to display the image
      */
-    double getWidth();
+    int getWidth();
 
     /**
      * Give the height of the area used to display the image (in pixels).
      *
      * @return the height of the area used to display the image
      */
-    double getHeight();
+    int getHeight();
 
     public boolean equals(Object o);
